+++ adbyss_psl/build.rs
@@ -33,8 +33,6 @@ type RawWildMap = HashMap<String, Vec<String>>;
 pub fn main() {
 	println!("cargo:rerun-if-env-changed=CARGO_PKG_VERSION");
 	println!("cargo:rerun-if-changed=skel/psl.rs.txt");
-	println!("cargo:rerun-if-changed=skel/raw/IdnaMappingTable.txt");
-	println!("cargo:rerun-if-changed=skel/raw/IdnaTestV2.txt");
 	println!("cargo:rerun-if-changed=skel/raw/public_suffix_list.dat");
 
 	// Sanity check. Obviously this won't change, but it is nice to know we